Use directories and separated image files to create the animations
Loading a animated gif is slow, and the activity does not reply
if a user press a few times the buttons.
To avoid that, we save the images in directories,
the dots used to display the advance, show the number of directories,
the buttons change to the next/previous animation,
and the timer load the next image available, or change to the next animation
if is the last image in the directory.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>